backend/requirements.txt
# YogaFlow Backend Dependencies
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0
python-multipart>=0.0.20

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
alembic>=1.14.0
greenlet>=3.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-dotenv>=1.0.1
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Validation & Data Processing
email-validator>=2.2.0
PyYAML>=6.0
Pillow>=10.0.0

# Email
aiosmtplib>=3.0.1
jinja2>=3.1.4

# Logging & Monitoring
structlog>=24.4.0
sentry-sdk[fastapi]>=2.0.0

# Production Database
asyncpg>=0.29.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
httpx>=0.28.0
